refactor: convert forms module into a package
We're going to have a lot of forms in this app, as medical information
is collected as part of the screening, and participant information such
as special appointment needs also needs additional forms. This module is
going to get huge if we don't split it up.
