+++ manage_breast_screening/mammograms/forms/ask_for_medical_information_form.py
@@ -1,0 +1,15 @@
+from django import forms
+
+
+class AskForMedicalInformationForm(forms.Form):
+    decision = forms.ChoiceField(
+        choices=(
+            ("yes", "Yes"),
+            ("no", "No - proceed to imaging"),
+        ),
+        required=True,
+        widget=forms.RadioSelect(),
+    )
+
+    def save(self):
+        pass
