+++ manage_breast_screening/mammograms/forms/record_medical_information_form.py
@@ -1,0 +1,15 @@
+from django import forms
+
+
+class RecordMedicalInformationForm(forms.Form):
+    decision = forms.ChoiceField(
+        choices=(
+            ("continue", "Yes, mark incomplete sections as ‘none’ or ‘no’"),
+            ("dropout", "No, screening cannot proceed"),
+        ),
+        required=True,
+        widget=forms.RadioSelect(),
+    )
+
+    def save(self):
+        pass
