+++ manage_breast_screening/mammograms/forms/screening_appointment_form.py
@@ -1,0 +1,15 @@
+from django import forms
+
+
+class ScreeningAppointmentForm(forms.Form):
+    decision = forms.ChoiceField(
+        choices=(
+            ("continue", "Yes, go to medical information"),
+            ("dropout", "No, screening cannot proceed"),
+        ),
+        required=True,
+        widget=forms.RadioSelect(),
+    )
+
+    def save(self):
+        pass
